Added reset button on nav bar and will only appear when results are on page
