+++ client/src/components/NavbarComponents/UnstyledNavbar.tsx
@@ -3,6 +3,11 @@ import styled from 'styled-components';
 //import MenuIcon from '@mui/icons-material/Menu';
 import {  DarkGreyGreen, LightGreyGreen,  } from '../../themes';
 import icon from "../../assets/icon.png";
+import questionsArr from '../../questionsLibrary';
+import { setQueryRes } from '../../Features/QueryResult/queryResultSlice';
+import { setCounter } from '../../Features/Questions/questionsCounter';
+import { useAppDispatch, useAppSelector } from '../../Hooks/hooks';
+import { RootState } from '../../App/store';
 
 const UnstyledNavContainer = styled.nav`
   display: flex;
@@ -48,8 +53,31 @@ const LeafIcon = styled.img`
     background-color:${DarkGreyGreen.secondary1.color};
   }
 `;
+const ResetBtn = styled.button`
+font-size: 1em;
+margin: 1em;
+border-radius: 5px;
+background-color: floralwhite;
+color: ${DarkGreyGreen.primary1.color};
+padding: 0.25em 1em;
+border: none;
+
+&:hover {
+  background-color: ${DarkGreyGreen.secondary1.color};
+}
+`;
 
 const UnstyledNavbar: React.FC = () => {
+  const queryResult = useAppSelector((state: RootState) => state.queryResult);
+  const dispatch = useAppDispatch();
+
+  const reset = () =>{
+    dispatch(setQueryRes([]));
+    dispatch(setCounter(0));
+    for(let i =1; i < questionsArr.length; i++){
+      questionsArr[i].isAnswered = false;
+    }
+  }
   return (
     <UnstyledNavContainer>
       <SpreadIcons>
@@ -58,9 +86,11 @@ const UnstyledNavbar: React.FC = () => {
             <LeafIcon src = {icon}/>
           </LeafColor>
         </a>
+        {queryResult.length > 0 && <ResetBtn onClick = {reset}>Restart Search</ResetBtn>}
       </SpreadIcons>
     </UnstyledNavContainer>
   )
 };
 
 export default UnstyledNavbar;
+
